% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitchisenVariation.R
\name{aitchisenVariation}
\alias{aitchisenVariation}
\title{Aitchisen's Variation Matrix}
\usage{
aitchisenVariation(X)
}
\arguments{
\item{X}{An n x p data matrix.}
}
\value{
A p x p Aitchisen variation matrix.
}
\description{
Computes Aitchisen's variation matrix from a matrix X
}
