% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitchisonVariation.R
\name{aitchisonVariation}
\alias{aitchisonVariation}
\title{Aitchison's Variation Matrix}
\usage{
aitchisonVariation(X)
}
\arguments{
\item{X}{An n x p data matrix.}
}
\value{
A p x p Aitchison variation matrix.
}
\description{
Computes Aitchison's variation matrix from a matrix X
}
