% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{biplot.KPR}
\alias{biplot.KPR}
\title{A "supervised" GMD biplot}
\usage{
\method{biplot}{KPR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{KPR}}

\item{...}{Additional parameters, used by the \code{inference} function call.}
}
\description{
This plots a GMD biplot of an object of class \code{KPR}. The axes correspond to first and second
GMD components of the data, as described by Wang et al. \code{biplot.KPR}
performs inference on the \code{KPR} object, and plots arrows corresponding to
penalized variables deemed significant. The points represent samples and are
colored with respect to the response vector \code{Y}.
}
\references{
Wang et al. Technical report.
}
