% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flax}
\alias{flax}
\title{Flaxseed lignan extract intervention microbiome data}
\format{A list of the four matrices descibed above.}
\usage{
data(flax)
}
\description{
Baseline microbiome data (X) from a study on the effects of flaxseed lignan extract intervention on enterolactone production (Y).
The H matrix is a similarity kernel generated from unweighted Unifrac distances between the samples. The Q matrix is a similarity kernel generated from an Aitchisen variation matrix computed between the p variables.
The Y vector is already centered, and the X matrix has already been normalized with the centered log ratio and then column-centered.
}
\examples{
data(flax)
KPR(X = flax$X, Y = flax$Y, Q = flax$Q, H = flax$H)
}
\references{
Lampe et al. (2019) The American Journal of Clinical Nutrition
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31175806}{PubMed})
}
\keyword{datasets}
