% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{GMD.biplot}
\alias{GMD.biplot}
\title{A "supervised" GMD biplot}
\usage{
GMD.biplot(X, Y = NULL, H = diag(nrow(X)), Q = diag(ncol(X)), K = 10)
}
\arguments{
\item{X}{An n by p data matrix.}

\item{Y}{An optional outcome vector. If provided, the points in the biplot will be colored according to the corresponding value in Y.}

\item{H}{An n by n sample-wise similarity kernel.}

\item{Q}{A p by p variable-wise similarity kernel.}

\item{K}{An integer, corresponding to the number of GMD components to compute.}
}
\description{
This plots a GMD biplot where the axes correspond to first and second
GMD components of the data, as described by Wang et al.
}
\references{
Wang et al. Technical report.
}
