% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{GMD}
\alias{GMD}
\title{Generalized Matrix Decomposition}
\usage{
GMD(X, H, Q, K)
}
\arguments{
\item{X}{An n x p data matrix.}

\item{H}{An n x n sample-wise similarity kernel.}

\item{Q}{A p x p variable-wise similarity kernel.}

\item{K}{The number of GMD components to include in the decomposition.}
}
\value{
A list of matrices involved in the decomposition.
}
\description{
Computes the generalized matrix decomposition of a matrix X with respect to H and Q.
}
