% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPR.R
\name{KPR}
\alias{KPR}
\title{Kernel Penalized Regression}
\usage{
KPR(
  designMatrix,
  covariates = NULL,
  Y,
  H = diag(nrow(designMatrix)),
  Q = diag(ncol(designMatrix)),
  scale = FALSE,
  REML = FALSE
)
}
\arguments{
\item{designMatrix}{An n x p data matrix, consisting of variables that should be penalized by \code{Q}. Should be scaled and centered.}

\item{covariates}{An n x r data matrix, consisting of variables that should not be penalized. Should be scaled and centered.}

\item{Y}{An n x 1 response vector. Should be scaled and centered.}

\item{H}{An n x n sample similarity kernel. Must be symmetric and positive semidefinite. This defaults to an identity matrix.}

\item{Q}{A p x p variable similarity kernel. Must be symmetric and postive semidefinite. This defaults to an identity matrix.}

\item{scale}{Logical, indicates whether to scale \code{Q}, \code{H} and the design matrix to have a spectral norm of 1.}

\item{REML}{Logical, indicates whether to use REML estimation for finding the parameters. This will only work with a single H and Q matrix.}
}
\value{
\item{beta.hat}{Estimated coefficients for the penalized variables.}
\item{eta.hat}{Estimated coefficients for the unpenalized variables.}
\item{theta.hat}{The vector of tuning parameter values found with Maximum Likelihood.}
}
\description{
Fits a kernel penalized regression model using a design matrix X, response vector Y, sample similarity kernel H, and variable similarity kernel Q.
}
\references{
Randolph et al. (2018) The Annals of Applied Statistics
(\href{https://projecteuclid.org/euclid.aoas/1520564483}{Project Euclid})
}
