% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yatsunenko}
\alias{yatsunenko}
\title{Age-related microbiome data}
\format{A list with the following entries:
\describe{
  \item{\code{raw.counts}}{The raw abundance counts for each sample, across 149 genera.}
  \item{\code{age}}{The age (in years) of each subject.}
  \item{\code{patristic}}{Patristic differences between each of the 149 genera.}
  \item{\code{unifrac}}{Unweighted UniFrac distances bewteen each of the 100 samples.}
  \item{\code{ec}}{Enzyme commission data; see the paper cited below for reference.}
  \item{\code{geography}}{Country of origin for each sample.}
}}
\usage{
data(yatsunenko)
}
\description{
Raw microbiome abundance data (n = 100, p = 149 genera) from the geographical study by Yatsunenko et al (2012).
}
\references{
Yatsunenko et al. (2012) Nature (\href{https://www.ncbi.nlm.nih.gov/pubmed/22699611}{PubMed})

See the KPR vignette for an in-depth example.
}
\keyword{datasets}
