% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference for Kernel Penalized Regression models}
\usage{
inference(KPR.output, method = "GMD", ...)
}
\arguments{
\item{KPR.output}{The output of running the \code{KPR} function.}

\item{method}{A string specifying the inference method to use.}

\item{...}{Additional parameters corresponding to specific inference methods.}
}
\description{
Implementation of various inference methods for high dimensional regression. Currently, the available methods are the GMD inference (used by default), the Grace inference method, and the ridge projection method, from the \code{hdi} package. To select the Grace method, set \code{method = "Grace"} and to select the ridge projection method, set \code{method = "hdi"}.
}
\details{
Before calling \code{ridge.proj}, we transform the data with a Cholesky decomposition, which allows us to formulate the model as a ridge problem while still including information about H and Q. However, this means that when a non-trivial Q is included in the KPR model, \code{ridge.proj} will return p-values corresponding to the DPCoA estimates, not the true KPR estimates. This may slightly affect interpretability.

The value of lambda used in all inference methods is determined in the KPR model fit.
}
\references{
Wang et al. Technical report.

Dezeure et al. (2014) Statistical Science (\href{https://arxiv.org/abs/1408.4026}{arXiv})

Zhao and Shojaie. (2016) Biometrics (\href{https://arxiv.org/abs/1506.08339}{arXiv})
}
