% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference for Kernel Penalized Regression models}
\usage{
inference(KPR.output, method = "GMD", ...)
}
\description{
Inference for Kernel Penalized Regression models
}
