% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference for Kernel Penalized Regression models}
\usage{
inference(KPR.output, method = "GMD", ...)
}
\arguments{
\item{KPR.output}{The output of running the \code{KPR} function.}

\item{method}{A string specifying the inference method to use.}
}
\description{
Implementation of various inference methods for high dimensional regression models. Currently only the GMD Inference is available.
}
