% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Running the GMD inference on a KPR model}
\usage{
inference(KPR.output, mu = 1, r = 0.05, weight = TRUE, scale = TRUE, ...)
}
\arguments{
\item{KPR.output}{Output from running the \code{KPR} function.}

\item{mu}{GMD inference parameter}

\item{r}{GMD inference parameter}

\item{weight}{Logical, indicates whether to include a penalty factor when computing the beta.glasso vector.}

\item{scale}{Logical, indicates whether to scale the design matrix with respect to the eigenvalues of the composite Q matrix.}
}
\value{
An object of classes KPR with the following fields added:
\item{p.values}{P-values for each penalized coefficient, resulting from the GMD inference.}
\item{bound}{The stochastic bound used to compute each p-value.}
\item{sigmaepsi.hat}{Variance component estimate used in the GMD inference.}
}
\description{
Running the GMD inference on a KPR model
}
