% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPR.R
\name{KPR}
\alias{KPR}
\title{Kernel Penalized Regression}
\usage{
KPR(X, Y, H = diag(nrow(X)), Q = diag(ncol(X)), n.lambda = 200,
  lambda, K = 5)
}
\arguments{
\item{X}{An n x p data matrix. Should be scaled and centered.}

\item{Y}{An n x 1 response vector. Should be scaled and centered.}

\item{H}{An n x n sample similarity kernel. Must be symmetric and positive semidefinite. This defaults to an identity matrix.}

\item{Q}{A p x p variable similarity kernel. Must be symmetric and postive semidefinite. This defaults to an identity matrix.}

\item{n.lambda}{The number of lambda values to test through the cross validation search. The values are generated internally.}

\item{lambda}{A vector of lambda values to test through cross validation. This will override the sequence generated with the n.lambda parameter.}

\item{K}{The number of folds in the cross validation search.}
}
\value{
\item{betahat}{A matrix of estimated coefficients, where each column corresponds to a different value of lambda.}
\item{lambda.1se}{The largest value of lambda within one standard error of lambda.min.}
\item{lambda.1se.index}{The index of the \code{lambda.1se} value in the \code{lambda} vector. This can be used to retrieve the estimated coefficients corresponding
to \code{lambda.1se} from the \code{betahat} matrix.}
}
\description{
Fits a kernel penalized regression model using a design matrix X, response vector Y, sample similarity kernel H, and variable similarity kernel Q.
The lambda parameter is found through k-folds cross validation.
}
