% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPR.R
\name{KPR}
\alias{KPR}
\title{Kernel Penalized Regression}
\usage{
KPR(
  X,
  E = NULL,
  Y,
  H = diag(nrow(X)),
  Q = diag(ncol(X)),
  scale = FALSE,
  REML = FALSE,
  Q.inv = TRUE,
  control.outer = list(trace = FALSE, NMinit = TRUE, method = "BFGS"),
  control.optim = list()
)
}
\arguments{
\item{X}{An n x p data matrix, consisting of variables that should be penalized by the Q matrices. Should be scaled and centered.}

\item{E}{An n x r data matrix, consisting of variables that should not be penalized. Should be scaled and centered.}

\item{Y}{An n x 1 response vector. Should be scaled and centered.}

\item{H}{A list of n x n sample similarity kernels. If only one matrix is included in the model, it does not need to be wrapped as a list. All matrices must be symmetric positive semidefinite. This defaults to a single identity matrix.}

\item{Q}{A list of p x p variable similarity kernels. If only one matrix is included in the model, it does not need to be wrapped as a list. All matrices must be symmetric positive semidefinite. This defaults to a single identity matrix.}

\item{scale}{Logical, indicates whether to scale all the Q's, H's and the design matrix to have a spectral norm of 1.}

\item{REML}{Logical, indicates whether to use REML estimation for finding the parameters. This will only work with a single H and Q matrix, and is the preferred method in this case.}

\item{Q.inv}{Logical, indicates whether to penalize Q_composite or Q_composite inverse.}

\item{control.outer}{A list of parameters used by the outer loop in `constrOptim.nl`. This is only used when `REML = FALSE`.}

\item{control.optim}{A list of parameters used by the inner loop in `constrOptim.nl`.}
}
\value{
\item{beta.hat}{Estimated coefficients for the penalized variables.}
\item{eta.hat}{Estimated coefficients for the unpenalized variables.}
\item{lambda}{The optimal lambda parameter estimated with maximum likelihood.}
\item{alpha}{The vector of optimal weights corresponding to the Q matrices.}
\item{sigma}{The vector of optimal weights corresponding to the H matrices.}
}
\description{
Fits a kernel penalized regression model using a design matrix X, response vector Y, sample similarity kernels H_1, H_2, ..., H_h, and variable similarity kernels Q_1, ..., Q_q.
}
\references{
Randolph et al. (2018) The Annals of Applied Statistics
(\href{https://projecteuclid.org/euclid.aoas/1520564483}{Project Euclid})
}
