% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPR.R
\name{KPR}
\alias{KPR}
\title{Kernel Penalized Regression}
\usage{
KPR(designMatrix, covariates, Y, H = diag(nrow(designMatrix)),
  Q = diag(ncol(designMatrix)), REML = TRUE, n.lambda = 200, lambda,
  K = 5, useCpp = TRUE, seed, scale = TRUE, inference = TRUE, ...)
}
\arguments{
\item{designMatrix}{An n x p data matrix, consisting of variables that should be penalized by \code{Q}. Should be scaled and centered.}

\item{covariates}{An n x r data matrix, consisting of variables that should not be penalized. Should be scaled and centered.}

\item{Y}{An n x 1 response vector. Should be scaled and centered.}

\item{H}{An n x n sample similarity kernel. Must be symmetric and positive semidefinite. This defaults to an identity matrix.}

\item{Q}{A p x p variable similarity kernel. Must be symmetric and postive semidefinite. This defaults to an identity matrix.}

\item{REML}{logical, indicates whether to use REML to find the optimal value of \code{lambda}. If a value of \code{lambda} is given, this is ignored.}

\item{n.lambda}{The number of lambda values to test through the cross validation search. The values are generated internally.}

\item{lambda}{A vector of lambda values to test through cross validation. This will override the sequence generated with the \code{n.lambda} parameter.}

\item{K}{The number of folds in the cross validation search.}

\item{useCpp}{Indicate whether to use the C++ backend for cross-validation. This is ignored if `REML` is set to TRUE.}

\item{seed}{Set a seed for random number generation.}

\item{scale}{Logical, indicates whether to scale \code{Q} and the design matrix.}

\item{inference}{Logical, indicates whether to compute p-values for penalized regression coefficients with the GMD inference.}

\item{...}{Additional parameters passed to the GMD inference}
}
\value{
\item{beta.hat}{A matrix of estimated coefficients for the penalized variables, where each column corresponds to a different value of lambda.}
\item{eta.hat}{A matrix of estimated coefficients for the unpenalized variables, where each column corresponds to a different value of lambda.}
\item{lambda}{The vector of lambda values used in cross validation.}
\item{lambda.min}{The value of lambda that results in the minimum sum of squared errors.}
\item{lambda.min.index}{The index of \code{lambda.min} in the \code{lambda} vector. This can be used to retrieve the estimated coefficients corresponding
to \code{lambda.min} from the \code{betahat} matrix.}
\item{lambda.1se}{The largest value of lambda within one standard error of \code{lambda.min}.}
\item{lambda.1se.index}{The index of the \code{lambda.1se} value in the \code{lambda} vector.}
\item{cv.errors}{The error matrix generated in the cross-validation procedure.}
\item{REML}{Was REML used to find \code{lambda}?}
\item{p.values}{P-values for each penalized coefficient, resulting from the GMD inference.}
\item{bound}{The stochastic bound used to compute each p-value.}
}
\description{
Fits a kernel penalized regression model using a design matrix X, response vector Y, sample similarity kernel H, and variable similarity kernel Q.
}
\references{
Randolph et al. (2018) The Annals of Applied Statistics
(\href{https://projecteuclid.org/euclid.aoas/1520564483}{Project Euclid})

Wang et al. Technical report.
}
