% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPR.R
\name{KPR}
\alias{KPR}
\title{Kernel Penalized Regression}
\usage{
KPR(
  designMatrix,
  covariates,
  Y,
  H = diag(nrow(designMatrix)),
  Q = diag(ncol(designMatrix)),
  lambda,
  scale = TRUE,
  inference = TRUE,
  ...
)
}
\arguments{
\item{designMatrix}{An n x p data matrix, consisting of variables that should be penalized by \code{Q}. Should be scaled and centered.}

\item{covariates}{An n x r data matrix, consisting of variables that should not be penalized. Should be scaled and centered.}

\item{Y}{An n x 1 response vector. Should be scaled and centered.}

\item{H}{An n x n sample similarity kernel. Must be symmetric and positive semidefinite. This defaults to an identity matrix.}

\item{Q}{A p x p variable similarity kernel. Must be symmetric and postive semidefinite. This defaults to an identity matrix.}

\item{lambda}{A fixed value of lambda to be used in fitting the KPR model. If this is missing, an appropriate value is chosen using REML estimation.}

\item{scale}{Logical, indicates whether to scale \code{Q} and the design matrix.}

\item{inference}{Logical, indicates whether to compute p-values for penalized regression coefficients with the GMD inference.}

\item{...}{Additional parameters passed to the GMD inference}
}
\value{
\item{beta.hat}{A matrix of estimated coefficients for the penalized variables, where each column corresponds to a different value of lambda.}
\item{eta.hat}{A matrix of estimated coefficients for the unpenalized variables, where each column corresponds to a different value of lambda.}
\item{lambda}{The vector of lambda values used in cross validation.}
\item{p.values}{P-values for each penalized coefficient, resulting from the GMD inference.}
\item{bound}{The stochastic bound used to compute each p-value.}
\item{sigmaepsi.hat}{Variance component estimate used in the GMD inference.}
}
\description{
Fits a kernel penalized regression model using a design matrix X, response vector Y, sample similarity kernel H, and variable similarity kernel Q.
}
\references{
Randolph et al. (2018) The Annals of Applied Statistics
(\href{https://projecteuclid.org/euclid.aoas/1520564483}{Project Euclid})

Wang et al. Technical report.
}
